+++ .gitignore
@@ -1,2 +1,3 @@
 .idea/
 *.xml
+.idea/workspace.xml
